Fix requested changes and invalidation
