+++ hooks/use-add-attendee-to-event.ts
@@ -34,6 +34,7 @@ export function useAddAttendeeToEvent(eventId: string) {
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ["attendees"] });
+      queryClient.invalidateQueries({ queryKey: ["events", eventId] });
     },
     onError: () => {
       showToastable({
