+++ hooks/use-attendance.ts
@@ -1,5 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
+import { usePayload } from "./use-payload";
 import { useStorage } from "./use-storage";
 
 import { PayloadConfig } from "@/utils/payload-client";
@@ -49,14 +50,41 @@ export function useAllAttendance() {
     queryKey: ["attendance"],
     queryFn: async () => {
       const keys = (await storage.getKeys()).filter((key) =>
-        key.startsWith(baseKey),
+        key.startsWith(baseKey)
       );
       return Promise.all(
         keys.map(async (key) => {
           const obj = await storage.getObject(key);
           return obj as PayloadConfig["collections"]["attendee"];
-        }),
+        })
       );
+    },
+  });
+}
+
+export function useCancelAttendance(eventId: string) {
+  const storage = useStorage();
+  const queryClient = useQueryClient();
+  const payload = usePayload();
+  const key = makeKey(eventId);
+
+  return useMutation({
+    mutationFn: async () => {
+      const attendeeName = (await storage.getObject(key)) as { name: string };
+      await storage.removeItem(key);
+
+      await payload.collections.attendee.delete({
+        query: {
+          event: eventId,
+        },
+        where: {
+          name: { equals: attendeeName.name },
+        },
+      });
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: [key] });
+      queryClient.invalidateQueries({ queryKey: ["events", eventId] });
     },
   });
 }
